9.14.0.0 support. Read description!
WARNING: as of 9.14.0.0, Snapchat started asserting a header value we
cannot generate on our own. There is a way around this but it involves
sending the username and password over an unencrypted network. You must
explicitly set the `useInsecureLogin` property of SKClient to YES
before signing in.

Also fixed some stories not downloading properly or not at all, and
added an OS X project.